Add integration test with real throughput data to verify correct mc result

+++ CMFTAspNet/Models/Forecast/WhenForecast.cs
@@ -21,7 +21,7 @@ namespace CMFTAspNet.Models.Forecast
                 }
             }
 
-            return 100 / totalTrials * trialCounter;
+            return 100 / ((double)totalTrials) * trialCounter;
         }
     }
 }
